add module structure
